Separated tests into data type and official tests, new integration tests for dist and CLI (commented out)
