+++ .travis.yml
@@ -21,4 +21,6 @@ matrix:
     - os: linux
       node_js: "6"
       env: TEST_SUITE=build
+before_script:
+  - npm run build
 script: npm run $TEST_SUITE
